+++ datamodel/odata-client/src/main/java/com/sap/cloud/sdk/datamodel/odata/client/request/ODataHealthyResponseValidator.java
@@ -4,8 +4,6 @@
 
 package com.sap.cloud.sdk.datamodel.odata.client.request;
 
-import java.nio.charset.StandardCharsets;
-
 import javax.annotation.Nonnull;
 
 import org.apache.http.HttpResponse;
@@ -22,6 +20,8 @@ import com.sap.cloud.sdk.result.GsonResultObject;
 import com.sap.cloud.sdk.result.ResultObject;
 
 import io.vavr.control.Try;
+
+import java.nio.charset.StandardCharsets;
 
 /**
  * Utility class to enable a healthy response validation.
@@ -46,23 +46,25 @@ class ODataHealthyResponseValidator
         final HttpResponse httpResponse = result.getHttpResponse();
         final StatusLine statusLine = httpResponse.getStatusLine();
 
+        // WIP
+        final Integer contentId =
+                httpResponse instanceof MultipartHttpResponse
+                        ? ((MultipartHttpResponse) httpResponse).getContentId()
+                        : null;
+        Integer failedBatchRequestNumber = null;
+        try
+        {
+            final byte[] content = httpResponse.getEntity().getContent().readAllBytes();
+            failedBatchRequestNumber =
+                    Integer.parseInt(new String(content, StandardCharsets.UTF_8).split("Content-ID: ")[1].split("\r\n")[0]);
+        } catch(Exception ignored) {
+        }
         if( statusLine != null && statusLine.getStatusCode() < HttpStatus.SC_BAD_REQUEST ) { // code < 400
             return;
         }
-
-        final Integer contentId =
-            httpResponse instanceof MultipartHttpResponse
-                ? ((MultipartHttpResponse) httpResponse).getContentId()
-                : null;
-
-        ODataRequestGeneric batchFailedRequest = null;
-        if( request instanceof ODataRequestBatch oDataRequestBatch ) {
-            batchFailedRequest = findFailedBatchRequest(httpResponse, oDataRequestBatch);
-        }
         final Integer statusCode = statusLine == null ? null : statusLine.getStatusCode();
         final String msg = "The HTTP response code (" + statusCode + ") indicates an error.";
-        final ODataResponseException preparedException =
-            new ODataResponseException(request, httpResponse, msg, null, batchFailedRequest);
+        final ODataResponseException preparedException = new ODataResponseException(request, httpResponse, msg, null);
 
         final Try<ODataServiceError> odataError = Try.of(() -> loadErrorFromResponse(result));
         if( odataError.isSuccess() ) {
@@ -71,33 +73,6 @@ class ODataHealthyResponseValidator
         }
 
         throw preparedException;
-    }
-
-    private static
-        ODataRequestGeneric
-        findFailedBatchRequest( final HttpResponse httpResponse, final ODataRequestBatch oDataRequestBatch )
-    {
-        try {
-            final byte[] content = httpResponse.getEntity().getContent().readAllBytes();
-            final int failedBatchRequestNumber =
-                Integer.parseInt(new String(content, StandardCharsets.UTF_8).split("Content-ID: ")[1].split("\r\n")[0]);
-            for( final ODataRequestBatch.BatchItem requestGeneric : oDataRequestBatch.getRequests() ) {
-                if( requestGeneric instanceof ODataRequestBatch.BatchItemChangeset changeset ) {
-                    for( final ODataRequestBatch.BatchItemSingle single : changeset.getRequests() ) {
-                        if( single.getContentId() == failedBatchRequestNumber ) {
-                            return single.getRequest();
-                        }
-                    }
-                } else if( requestGeneric instanceof ODataRequestBatch.BatchItemSingle single ) {
-                    if( single.getContentId() == failedBatchRequestNumber ) {
-                        return single.getRequest();
-                    }
-                }
-            }
-        }
-        catch( final Exception ignored ) {
-        }
-        return null;
     }
 
     @Nonnull